refactor(api): extract canvas-suggest prompt + LLM helpers
Lifts the prompt builder and LLM-call retry/validation logic out of
aiSuggestRoute.ts into reusable services so the upcoming streaming
chat-edit controller can share the exact same prompt + worker pool
plumbing without duplicating ~370 lines.

- buildCanvasSuggestPrompt: system-prompt builder, with optional
  contextHints { citations, prose } that appends a RECHERCHE-KONTEXT
  section (capped at 5 citations + 500-char prose excerpt) so callers
  can seed the model with research findings.
- runCanvasSuggest: tool-schema construction, two-attempt retry,
  tool-call extraction (canonical + Anthropic shapes), schema
  validation, supported-op filtering. Returns Result<{suggestions} | {error}>.

aiSuggestRoute.ts becomes a one-line delegate (423 → 52 LOC). No
behavior change — same model, temperature, tool, retry policy.
